Extract shared component-kit host vite plugins; serve kits on the addin linked dev server
